Fix Android logging
