+++ Cargo.toml
@@ -18,13 +18,10 @@ glam = { version = "0.30", features = [
 ] } #SIMD supported vector types\
 palette = "0.7.6"
 cfg-if = "1"
-tracing = { version = "0.1", features = [
-    "std",
-], default-features = false }
+tracing = { version = "0.1", features = ["std"], default-features = false }
 tracing-subscriber = { version = "0.3", default-features = false }
-tracing-error = { version = "0.1", default-features = false }
-paranoid-android = { version = "0.1.2" }
-thiserror = "1"
+tracing-error = { version = "*", default-features = false }
+thiserror = "2"
 
 # for reading and writing files
 serde_json = { version = "1.0", optional = true }
@@ -40,8 +37,10 @@ tracing-subscriber = { version = "0.3", features = [
 ], default-features = false }
 
 [target.'cfg(target_os = "android")'.dependencies]
-android_logger = "0.13"
-log = "0.4.25"
+android_logger = "0.14"
+log = "0.4"
+paranoid-android = { version = "*", optional = true }
+
 
 [build-dependencies]
 cbindgen = "0.28"
